backend/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
requests==2.32.5
simple-websocket==1.1.0
SQLAlchemy==2.0.48
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wsproto==1.3.2
